Airlines Seating Chart - Program 1 CSC 3000
This program takes a data file that has multiple passengers and their flight info. The program reads in the data and seats the passengers accordingly. It is first come first serve for seats. If the seat is taken then the row is checked for an empty seat. If the row is full the selected column is checked. If the selected column is full, the program checks the selected section and starts at the first seat and places the passenger in the next empty seat. If all seats are full in the passengers selected section, then they are booted to the waiting list.